Add clamWinSigUpdate script
Script designed to update the "unofficial" signatures of ClamAV which are provided by third parties such as SaneSecurity.

This script is designed for those using ClamWin Portable. Notes are provided in the script on which lines of code to modify for those using a permanent installation of ClamWin and/or those who are not interested in manually copying the files after each run.